Relocate runFrameworkGenerator
