+++ packages/create-cloudflare/src/helpers/command.ts
@@ -1,11 +1,7 @@
-import { logRaw, stripAnsi, updateStatus } from "@cloudflare/cli";
-import { dim } from "@cloudflare/cli/colors";
+import { stripAnsi } from "@cloudflare/cli";
 import { isInteractive, spinner } from "@cloudflare/cli/interactive";
 import { spawn } from "cross-spawn";
-import { getFrameworkCli } from "frameworks/index";
 import { quoteShellArgs } from "../common";
-import { detectPackageManager } from "./packageManagers";
-import type { C3Context } from "types";
 
 /**
  * Command is a string array, like ['git', 'commit', '-m', '"Initial commit"']
@@ -162,36 +158,4 @@ export const printAsyncStatus = async <T>({
 	}
 
 	return promise;
-};
-
-/**
- * Run a scaffolding tool with `npx` or its equivalent. The `ctx` object must be
- * populated with a framework that exists `src/frameworks/package.json`.
- *
- * @param ctx - The C3 context object
- * @param args - An array of additional arguments to be used
- */
-export const runFrameworkGenerator = async (ctx: C3Context, args: string[]) => {
-	const cli = getFrameworkCli(ctx, true);
-	const { npm, dlx } = detectPackageManager();
-	// yarn cannot `yarn create@some-version` and doesn't have an npx equivalent
-	// So to retain the ability to lock versions we run it with `npx` and spoof
-	// the user agent so scaffolding tools treat the invocation like yarn
-	const cmd = [...(npm === "yarn" ? ["npx"] : dlx), cli, ...args];
-	const env = npm === "yarn" ? { npm_config_user_agent: "yarn" } : {};
-
-	if (ctx.args.additionalArgs?.length) {
-		cmd.push(...ctx.args.additionalArgs);
-	}
-
-	updateStatus(
-		`Continue with ${ctx.template.displayName} ${dim(
-			`via \`${quoteShellArgs(cmd)}\``
-		)}`
-	);
-
-	// newline
-	logRaw("");
-
-	await runCommand(cmd, { env });
 };
