+++ packages/create-cloudflare/templates/hono/c3.ts
@@ -1,5 +1,5 @@
 import { logRaw } from "@cloudflare/cli";
-import { runFrameworkGenerator } from "helpers/command";
+import { runFrameworkGenerator } from "frameworks/index";
 import type { TemplateConfig } from "../../src/templates";
 import type { C3Context } from "types";
